refactor: update translations, cv
