+++ i18n/locales/en-GB.json
@@ -27,6 +27,8 @@
   "location_title": "Available on-site or remote, wherever you are :)",
   "view": "View",
   "contact": "Contact me",
+  "last_name": "Name",
+  "first_name": "",
   "contact_text": "and let's discuss your project.",
   "desc": "Describe your project...",
   "send": "Send",
